设备信息表初稿，码牌相关修改和优化

+++ aspnet-core/src/AGooday.AgPay.Application/Interfaces/IQrCodeShellService.cs
@@ -5,9 +5,9 @@ namespace AGooday.AgPay.Application.Interfaces
     public interface IQrCodeShellService : IDisposable
     {
         bool Add(QrCodeShellDto dto);
-        bool Remove(int recordId);
+        bool Remove(long recordId);
         bool Update(QrCodeShellDto dto);
-        QrCodeShellDto GetById(int recordId);
+        QrCodeShellDto GetById(long recordId);
         IEnumerable<QrCodeShellDto> GetAll();
         PaginatedList<T> GetPaginatedData<T>(QrCodeShellQueryDto dto);
     }
